+++ src/support/hash.h
@@ -35,7 +35,7 @@ inline uint32_t rehash(uint32_t x, uint32_t y) { // see http://www.cse.yorku.ca/
   return hash;
 }
 
-inline size_t rehash(uint64_t x, uint64_t y) { // see boost and https://stackoverflow.com/a/2595226/1176841
+inline uint64_t rehash(uint64_t x, uint64_t y) { // see boost and https://stackoverflow.com/a/2595226/1176841
   return x ^ (y + 0x9e3779b9 + (x << 6) + (x >> 2));
 }
